Update player before applying inputs
On the client
=============

We were updating the inputs of our player without calculating the
up-to-date version of its properties first, and as a consequence, during the
next logic step, we would update the player as if it had those inputs
since the last logic step.

As a result, there is a range of time in which we update the player
incorrectly using the new inputs instead of the previous ones.

On the server
=============

When receiving a 'move' event, we were updating the player's inputs and
its timestamp without calculating the up-to-date version of its properties
first. As a result, the player would not update for some time,
specifically since the last logic step til the received event time.